ChatBot - Clothe's Store/requirements.txt.txt
openai>=1.2.0
gradio>=3.50.0
python-dotenv>=1.0.0
